Rename DLLEXPORT_drakeCollision to drakeCollisionMacros
