+++ systems/plants/collision/Element.h
@@ -8,7 +8,7 @@
 #include <Eigen/Dense>
 #include <Eigen/StdVector>
 
-#include "DLLEXPORT_drakeCollision.h"
+#include "drakeCollisionMacros.h"
 #include "Geometry.h"
 
 namespace DrakeCollision
